drone-motion/src/requirements.txt
Pillow
termcolor
ruamel.yaml
